Perform flush before loading initial state.
This ensures MongoChangeStreamReceiver always sees at least one change
stream event that preceeds the initial state, and so
MongoChangeStreamReceiver.reconnectCursor always has a valid resume
token to use for reconnecting.

There are still some really unusual initialization-time error scenarios
in which we might attempt to reconnect without having seen a resume
token. In that case, we're a bit hosed. In that case, we report an error
and do our best to leave the application functioning.