MongoDriverSpecialTest and DisruptsMongoService.
Make a simple conformance test called MongoDriverConformanceTest, and
put the other MongoDB-specific tests in MongoDriverSpecialTest.

For efficiency, we want to boot just one MongoDB container, and
Toxiproxy only supports one proxy per port, hence we're only using one
proxy too. That means that tests that make use of Toxiproxy will
interfere with other tests, and can therefore not be run at the same
time. Hence, we add annotations UsesMongoService and DisruptsMongoService,
with the rule that only one @DisruptsMongoService can run at a time, and
it can't run at the same time as any @UsesMongoService.

I'm not thrilled with having to use these annotations. I would much
rather have had a separate Toxiproxy for each test, but it seems
Toxiproxy doesn't support that.

I've opened a PR to add support. Once this PR (or something similar) is
merged, we can remove the annotations and run the tests in parallel.

  https://github.com/testcontainers/testcontainers-java/pull/5506