Clean up BoskDriver exceptions.
This is system-level software. It's important that it's reliable.
We can't just throw RuntimeException everywhere without eroding the
user's ability to construct a system that is reliable in the presence
of errors.

However it's also possible to go too far. Any substrate in which Bosk is
running likely has a reasonable approach to handling runtime errors
already, and in such a world, checked exceptions are just an annoyance.

As a compromise, we declare a couple of exceptions that extend IOException:
InitializationFailureException and FlushFailureException. We feel these
bring the appropriate amount of attention to the fact that these methods
can fail at runtime, without burdening systems that already have a
straightforward way to deal with IOException.

Note that InterruptedException is declared separately, because it
requires catch clauses to do special handling. A catch clause for
IOException is unlikely to be suitable for InterruptedException.

Also note that BoskDriver.initialRoot throws InvalidTypeException.
This is primarily a convenience because many initialRoot implementations
will create References, but it also hints at our intention that
InitializationFailureException be used for runtime failures rather than
coding errors. Code that could not work under any circumstances really
ought to throw something like InvalidTypeException,
IllegalArgumentException, UnsupportedOperationException, or even
AssertionError.