Delete TodoBoskTest.
I'd like to say a few words about TodoBoskTest.

This was the original home of the bosk library. It began in here, in the
Ice Cube repository, as an experiment. And now it has grown into its own
project, with a proper test suite and there's no longer any reason to
keep this test.

It served us well, and we wish it all the best in retirement.