Add eventDelayMS setting to help reproduce races.
This would probably be better done using Toxiproxy rather than changing
the library code.

Also, it would be best not to need a real-time delay to reproduce these
problems. A lazy mode that processes events as late as possible would be
better, but it's difficult to do that without accidentally fixing race
conditions caused by doing only a portion of the processing of one event.