Fix bug in initialRoot.
Code inspection revealed a race condition, which I provoked using
eventDelayMS. The fix is to null out taskRef when initialRoot is
actually finished; not when the task itself (which runs on a different
thread) finishes.

This might not have been a bug when I first wrote it, and the task was
running in the foreground thread, but it became a bug when I moved the
task to the background thread.