Create manifest file.
The goal here is to make sure that bosk will eventually disconnect
from a database whose format it does not recognize, rather than continue
to interfere with it indefinitely.

The recommendation for ops would still be to ensure older servers are
shut down before refurbishing to a new format that older servers don't
support.  The logic added in this PR merely reduces the risk window if
you fail to do so; it does not eliminate the risk entirely.  It's still
possible for an older server to perform a write before it notices (via a
change stream event) that the format changed.

Because the manifest document is new, we currently tolerate the case
where it does not yet exist, defaulting to the Sequoia format. Once all
production systems are refurbished, we could remove this default logic.