Revamp transactions in MainDriver.
Transactions don't work the way I assumed, so I've decided to make a
wrapper class, TransactionalCollection, that works the way I want.

It maintains a ThreadLocal<ClientSession> and injects that into every
operation, ensuring that we won't accidentally do a one-off operation
that punches through the transaction.

It's also flexible about transaction semantics. Code that needs its
operations to be in a single transaction can call
ensureTransactionStarted, and it's ok for this to be called repeatedly
at mutiple levels of the stack. Then MainDriver calls
commitTransactionIfAny at the end of every session (unless an exception
is thrown). The result is that transactions can be begun at any level of
the stack, and will be committed at the appropriate time.

This flexibility allows code like SequoiaDriver to remain agnostic about
transactions, since that driver doesn't actually need them for its own
sake; but MainDriver needs a transaction for refurbish, and it can
establish one with no cooperation from SequoiaDriver. Conversely,
PandoDriver will need transactions for almost every operation, and it
can establish them too; and yet later, if PandoDriver finds ways to skip
transactions in certain cases, it has the freedom to do so.