Fix: Use Replace for manifest.
A while back* we stopped deleting the manifest document during
refurbish. The reason was that if we delete the manifest on a refurbish,
it's pretty much inevitable that every FormatDriver would need to
disconnect and reconnect, even if it were not otherwise necessary,
because the database would temporarily be in a state where the
FormatDriver cannot determine whether its local state is correct
anymore. Even a FormatDriver refurbishing to its own format, which
should not cause a disruption at all, would cause an otherwise
unnecessary reconnection cycle.

One consequence of this change was that the UPDATE operation used to
change from, say, sequoia to pando format, would simply write to the
manifest's "pando" field, leaving the manifest document in an invalid
state where it has format configurations for both sequoia and pando.

This change fixes this by using a REPLACE instead of an UPDATE.

* Commit f40bfad07ca1390b4e2c7955f226ed6a4ccc9854