Fix: double-check that the root doc exists.
Pando's database writes use document filters to avoid doing writes under
two circumstances:

1. Conditional updates whose preconditions fail
2. Updates inside nonexistent state tree nodes

In either case, bosk is supposed to ignore the updates.

However, this led to a bug when refurbishing from pando to another
format. Refurbishing deletes any existing state documents, and this led
to a window after the delete but before the DELETE event arrives, during
which updates would be silently ignored.

As a workaround, when we detect that an update affected no documents, we
double check that the root document does, in fact, exist.

For Sequoia, for simplicity, we don't do this.